Ignore keyword arguments to Resp.read()
This commit adds a read() method to the Resp class,
which simply drops any keyword arguments that are sent
and calls the superclass method.

This restores compatibility with requests >= 1.2.1,
which sends a keyword argument without checking to see
if it's supported, therefore breaking compatibility
with most file-like objects.

ambv/requests-testadapter#1